docs: baseline-reference: clearer version info
Make it clear that the kernel versions 6.1 and 6.6 really refer to the
whole LTS branch, and that patch releases of Rust and Go are also
permitted by the respective baseline.

Also move the explanation for the :wrench: symbol to a separate
sentence, because it also appears in the text below, as opposed to just
the version table.