docs(newsletter): this-week-in-loongarch-16: improve readability after removal of erhua mark
In “对此对指令”, the two 对's are serving different grammatical roles:
the former is an explicit topic marker -- "as for", while the latter is
"pair". Formerly an erhua mark is placed after the latter for visually
hinting the reader, but after #113 we no longer tend to utilize erhua in
this way. Change to alternative expression that doesn't involve
"对"/"pair" to work around this.