docs(newsletter): this-week-in-loongarch-31: fix a grammatical ambiguity
the intended grouping was "((((大量)涉及)(浮点比较操作)的)函数)"
("functions that involves a lot of FP comparison operations"),
but on second thought it is not the obvious/trivial paraphrasing.
It is almost certainly going to be misunderstood as
"(大量(((涉及)(浮点比较操作))的)函数)" ("many functions that involve FP
comparisons") which is wrong.

Fix by adding more helper words to fixate the phrase structure.