feat: remove all trademark symbols
Trademark symbols are meant for use by the *trademark owners*, to make
the trademarked content more prominently so, *not us*. We as
third-parties disseminating neutral content about the respective
trademarked products, have no obligation to highlight the trademarks for
them -- we're unaffiliated after all!

Switch to a short, neutral (that doesn't single out Loongson; we're
discussing multiple other companies' products for a while already), and
hopefully correct notice in the page footer, and remove every "&reg;"
from the whole site.