fix(asmdb): fix some decodetree leaves not being really decided
For example the `asrt{le,gt}` insns are already unique given their
prefixes, but they additionally require the D slot to be all zeroes to
correctly decode. So the previous end condition for the decodetree
construction is not enough: we need to ensure all of the insn's fixed
bits are covered by previous look-at's, in addition to the match
category containing just one insn.