fix(asmdb): fix the /asmdb page's SSR behavior
The page loads fine when one navigates to it after the page is fully
loaded, but not when one accesses it directly (e.g. by opening it from a
bookmark or a link shared via IM), signifying an SSR rehydration problem.

Make the ThemeAwareAntdContainer component BrowserOnly to automatically
fix all pages making use of antd, as we expect all antd usage to respect
the Docusaurus color scheme.